Reorg of repo
So, one can now start the app using the command line (caveats below).
Atom can't do this yet. Haven't tried IntelliJ. The caveat is that the
app never gets off the ground because of some issue validating the Atom
installation (which it always does, even from the command line). This
may or may not be local to my installation.  But things are far enough
along I figured I'd put it back.